+++ modules/personal.nix
@@ -1,10 +1,6 @@
 {
   imports = [ "contacts" ];
 
-  nixosModule = { ... }: {
-    services.tailscale.enable = true;
-  };
-
   hmModule = { pkgs, ... }: {
     home.packages = builtins.attrValues {
       inherit (pkgs) discord joplin-desktop signal-desktop;